Update programme 3.py

+++ user-input/programme 3.py
@@ -2,14 +2,15 @@
 #Date = 1st of Semptember 2025
 #description = Program to request user name as input and return the same with a greeting 
 First_number = input(' first Number : ')
-First_number = int(First_number)
+First_number = float(First_number)
 
 Second_number = input('Second Number : ')
-Second_number = int(Second_number)
+Second_number = float(Second_number)
 third_number = input('Third number :  ')
-third_number = int(third_number)
+third_number = float(third_number)
 
 total2 = First_number + Second_number + third_number
 total3 = total + total2
 
 print(total3)
+
